requirements.txt
Flask==3.1.0
flask-cors==5.0.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
cryptography==44.0.0
gunicorn==23.0.0
python-dotenv==1.0.1
requests==2.32.3
